+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Changed default timezone offset to UTC
 - Added dictionary attribute with plotting settings to subscription
 - Added boolean attribute mute_new_metrics to trigger
+- Optimized tag stats methods (stats, fetch_assigned_triggers, fetch_assigned_triggers_by_tags)<br/>
+to return trigger id's instead of triggers to speed up method execution time
 
 # 2.3.5
 - Fixed url to update existing subscription